ref(browser): Add protocol metadata to resource spans
- Relocated `extractNetworkProtocol` util to `browser-utils`.
- Modified `browserMetrics` test to assert the new metadata.
