+++ packages/browser-utils/test/browser/browserMetrics.test.ts
@@ -131,6 +131,7 @@ describe('_addResourceSpans', () => {
       encodedBodySize: 256,
       decodedBodySize: 256,
       renderBlockingStatus: 'non-blocking',
+      nextHopProtocol: 'http/1.1',
     });
     _addResourceSpans(span, entry, resourceEntryName, 123, 456, 100);
 
@@ -150,6 +151,7 @@ describe('_addResourceSpans', () => {
       encodedBodySize: 256,
       decodedBodySize: 256,
       renderBlockingStatus: 'non-blocking',
+      nextHopProtocol: 'http/1.1',
     });
     _addResourceSpans(span, entry, 'https://example.com/assets/to/me', 123, 456, 100);
 
@@ -169,6 +171,7 @@ describe('_addResourceSpans', () => {
       encodedBodySize: 456,
       decodedBodySize: 593,
       renderBlockingStatus: 'non-blocking',
+      nextHopProtocol: 'http/1.1',
     });
 
     const timeOrigin = 100;
@@ -195,6 +198,8 @@ describe('_addResourceSpans', () => {
           ['url.scheme']: 'https',
           ['server.address']: 'example.com',
           ['url.same_origin']: true,
+          ['network.protocol.name']: 'http',
+          ['network.protocol.version']: '1.1',
         },
       }),
     );
@@ -233,6 +238,7 @@ describe('_addResourceSpans', () => {
       const { initiatorType, op } = table[i]!;
       const entry = mockPerformanceResourceTiming({
         initiatorType,
+        nextHopProtocol: 'http/1.1',
       });
       _addResourceSpans(span, entry, 'https://example.com/assets/to/me', 123, 234, 465);
 
@@ -254,6 +260,7 @@ describe('_addResourceSpans', () => {
       encodedBodySize: 0,
       decodedBodySize: 0,
       renderBlockingStatus: 'non-blocking',
+      nextHopProtocol: 'h2',
     });
 
     _addResourceSpans(span, entry, resourceEntryName, 100, 23, 345);
@@ -271,6 +278,8 @@ describe('_addResourceSpans', () => {
           ['url.scheme']: 'https',
           ['server.address']: 'example.com',
           ['url.same_origin']: true,
+          ['network.protocol.name']: 'http',
+          ['network.protocol.version']: '2',
         },
       }),
     );
@@ -288,6 +297,7 @@ describe('_addResourceSpans', () => {
       transferSize: 2147483647,
       encodedBodySize: 2147483647,
       decodedBodySize: 2147483647,
+      nextHopProtocol: 'h3',
     });
 
     _addResourceSpans(span, entry, resourceEntryName, 100, 23, 345);
@@ -301,6 +311,8 @@ describe('_addResourceSpans', () => {
           'server.address': 'example.com',
           'url.same_origin': true,
           'url.scheme': 'https',
+          ['network.protocol.name']: 'http',
+          ['network.protocol.version']: '3',
         },
         description: '/assets/to/css',
         timestamp: 468,
@@ -325,6 +337,7 @@ describe('_addResourceSpans', () => {
       transferSize: null,
       encodedBodySize: null,
       decodedBodySize: null,
+      nextHopProtocol: 'h3',
     } as unknown as PerformanceResourceTiming;
 
     _addResourceSpans(span, entry, resourceEntryName, 100, 23, 345);
@@ -338,6 +351,8 @@ describe('_addResourceSpans', () => {
           'server.address': 'example.com',
           'url.same_origin': true,
           'url.scheme': 'https',
+          ['network.protocol.name']: 'http',
+          ['network.protocol.version']: '3',
         },
         description: '/assets/to/css',
         timestamp: 468,
@@ -365,6 +380,7 @@ describe('_addResourceSpans', () => {
         encodedBodySize: 0,
         decodedBodySize: 0,
         deliveryType,
+        nextHopProtocol: 'h3',
       });
 
       _addResourceSpans(span, entry, resourceEntryName, 100, 23, 345);
